fix(test): extract classifyMesh into pure module so bun tests don't hang
test/car-wireframe.test.ts imported classifyMesh from CarBody.tsx, which
pulls @react-three/drei at module load. Drei's side-effect imports block
under Bun's test loader (no DOM / WebGL context) — the test runner hung
indefinitely on that file.

Moves classifyMesh + DEFAULT_HIDDEN_MESHES into wireframe/classify-mesh.ts
(no Three / drei imports). CarBody.tsx re-exports both symbols so
production callers still work unchanged. Test imports the pure module
directly.
